Js fonctionnel avec le webpack
