+++ templates/base.html.twig
@@ -11,7 +11,7 @@
               href="https://unpkg.com/leaflet.markercluster@1.3.0/dist/MarkerCluster.css" />
         <link rel="stylesheet" type="text/css"
               href="https://unpkg.com/leaflet.markercluster@1.3.0/dist/MarkerCluster.Default.css" />
-        <link rel="stylesheet" href="{{ asset('./../build/app.css') }}">
+        {{ encore_entry_link_tags('app') }}
     {% endblock %}
 </head>
 <body>
@@ -21,9 +21,7 @@
             crossorigin=""></script>
     <script type='text/javascript'
             src='https://unpkg.com/leaflet.markercluster@1.3.0/dist/leaflet.markercluster.js'></script>
-    <script type="text/javascript"
-            src="{{ asset('./../build/app.js') }}"></script>
-    {% include 'js.html.twig' %}
+    {{ encore_entry_script_tags('app') }}
 {% endblock %}
 </body>
 </html>
